refresh_token grant doesn't require client creds
This is a quick hack to allow for clients that are a hybrid of
confidential and public, where a trusted backend is used for the
authorization code flow while the refresh token (and access token(s))
are stored in a public client (single-page app, CLI, etc).

The proper way to address this would be for the oauth2 library to have
first-class support for public clients, including PKCE. This commit
instead accepts a slight hit to the security posture while making it
possible for such hybrid clients to use the refresh token flow.